backend/requirements.txt
fastapi
uvicorn[standard]
web3
python-dotenv
psycopg2-binary